Renames components and adds a new sample component. Also refactors page code to be outside of components directory
