+++ templates/vue/_files/app/vue/components/MainApp.vue
